# src/main/resources/openapi/wms_v1.yaml
openapi: 3.0.3
info:
  title: Warehouse Management System API
  description: API for managing items, orders, locations, and customers in a warehouse.
  version: 1.0.0
servers:
  - url: https://api.warehouse.com/v1
    description: Production server

paths:
  /item:
    get:
      summary: List all items
      description: Retrieve a list of all items in the warehouse.
      tags:
        - itemApi
      operationId: listItem
      responses:
        '200':
          description: A list of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
    post:
      summary: Create a new item
      description: Add a new item to the warehouse inventory.
      tags:
        - itemApi
      operationId: createItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCreate'
      responses:
        '201':
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'

  /item/{itemId}:
    get:
      summary: Retrieve item by ID
      description: Retrieve a specific item by its ID.
      tags:
        - itemApi
      operationId: retrieveItem
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A single item
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Item not found

    put:
      summary: Update item by ID
      description: Modify the details of a specific item entirely.
      tags:
        - itemApi
      operationId: updateItem
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemUpdate'
      responses:
        '200':
          description: Item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Item not found

    delete:
      summary: Delete item by ID
      description: Remove a specific item from the warehouse inventory.
      tags:
        - itemApi
      operationId: deleteItem
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Item deleted successfully
        '404':
          description: Item not found

  /order:
    get:
      summary: List all orders
      description: Retrieve a list of all orders.
      tags:
        - orderApi
      operationId: listOrder
      responses:
        '200':
          description: A list of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'

    post:
      summary: Create a new order
      description: Add a new order to the system.
      tags:
        - orderApi
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'

  /order/{orderId}:
    get:
      summary: Retrieve order by ID
      description: Retrieve details of a specific order.
      tags:
        - orderApi
      operationId: retrieveOrder
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A single order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found

    put:
      summary: Update order by ID
      description: Modify the details of a specific order entirely.
      tags:
        - orderApi
      operationId: updateOrder
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdate'
      responses:
        '200':
          description: Order updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found

    delete:
      summary: Delete order by ID
      description: Remove a specific order.
      tags:
        - orderApi
      operationId: deleteOrder
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Order deleted successfully
        '404':
          description: Order not found

  /location:
    get:
      summary: List all locations
      description: Retrieve a list of all storage locations in the warehouse.
      tags:
        - locationApi
      operationId: listLocation
      responses:
        '200':
          description: A list of locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Location'

    post:
      summary: Create a new location
      description: Add a new storage location.
      tags:
        - locationApi
      operationId: createLocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationCreate'
      responses:
        '201':
          description: Location created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'

  /location/{locationId}:
    get:
      summary: Retrieve location by ID
      description: Retrieve details of a specific storage location.
      tags:
        - locationApi
      operationId: retrieveLocation
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A single location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '404':
          description: Location not found

    put:
      summary: Update location by ID
      description: Modify the details of a specific location entirely.
      tags:
        - locationApi
      operationId: updateLocation
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LocationUpdate'
      responses:
        '200':
          description: Location updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '404':
          description: Location not found

    delete:
      summary: Delete location by ID
      description: Remove a specific storage location.
      tags:
        - locationApi
      operationId: deleteLocation
      parameters:
        - name: locationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Location deleted successfully
        '404':
          description: Location not found

  /customer:
    get:
      summary: List all customers
      description: Retrieve a list of all customers.
      tags:
        - customerApi
      operationId: listCustomer
      responses:
        '200':
          description: A list of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'

    post:
      summary: Create a new customer
      description: Add a new customer to the system.
      tags:
        - customerApi
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreate'
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'

  /customer/{customerId}:
    get:
      summary: Retrieve customer by ID
      description: Retrieve details of a specific customer.
      tags:
        - customerApi
      operationId: retrieveCustomer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A single customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found

    put:
      summary: Update customer by ID
      description: Modify the details of a specific customer entirely.
      tags:
        - customerApi
      operationId: updateCustomer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdate'
      responses:
        '200':
          description: Customer updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found

    delete:
      summary: Delete customer by ID
      description: Remove a specific customer.
      tags:
        - customerApi
      operationId: deleteCustomer
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Customer deleted successfully
        '404':
          description: Customer not found

components:
  schemas:
    Item:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the item.
        name:
          type: string
          description: Name of the item.
        quantity:
          type: integer
          description: Number of items available.
        description:
          type: string
        unitOfMeasure:
          type: string
        locationId:
          type: string
          description: ID of the storage location.
        price:
          type: number
          description: Price of the item.
      required:
        - id
        - name
        - quantity
        - unitOfMeasure
        - locationId
        - price

    ItemCreate:
      type: object
      properties:
        name:
          type: string
          description: Name of the item.
        quantity:
          type: integer
          description: Number of items available.
        locationId:
          type: string
          description: ID of the storage location.
        price:
          type: number
          description: Price of the item.
      required:
        - name
        - quantity
        - unitOfMeasure
        - locationId
        - price

    ItemUpdate:
      type: object
      properties:
        name:
          type: string
          description: Name of the item.
        quantity:
          type: integer
          description: Number of items available.
        locationId:
          type: string
          description: ID of the storage location.
        price:
          type: number
          description: Price of the item.
      required:
        - name
        - quantity
        - unitOfMeasure
        - locationId
        - price

    Order:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the order.
        customerId:
          type: string
          description: ID of the customer who placed the order.
        items:
          type: array
          description: List of ordered items.
          items:
            $ref: '#/components/schemas/Item'
        totalPrice:
          type: number
          description: Total price of the order.
        createdAt:
          type: string
          format: date-time
          description: Date when the order was placed.

    OrderCreate:
      type: object
      properties:
        customerId:
          type: string
          description: ID of the customer who placed the order.
        items:
          type: array
          description: List of ordered items.
          items:
            type: object
            properties:
              item:
                $ref: '#/components/schemas/Item'
              quantity:
                type: integer
            required:
              - item
              - quantity
      required:
        - items

    OrderUpdate:
      type: object
      properties:
        customerId:
          type: string
          description: ID of the customer who placed the order.
        items:
          type: array
          description: List of ordered items.
          items:
            type: object
            properties:
              item:
                $ref: '#/components/schemas/Item'
              quantity:
                type: integer
            required:
              - item
              - quantity
      required:
        - items

    Location:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the location.
        name:
          type: string
          description: Name of the storage location.
        capacity:
          type: integer
          description: Maximum capacity of the location.
        currentCapacity:
          type: integer
          description: Current capacity of the location.

    LocationCreate:
      type: object
      properties:
        name:
          type: string
          description: Name of the storage location.
        capacity:
          type: integer
          description: Maximum capacity of the location.

    LocationUpdate:
      type: object
      properties:
        name:
          type: string
          description: Name of the storage location.
        capacity:
          type: integer
          description: Maximum capacity of the location.

    Customer:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the customer.
        firstName:
          type: string
          description: Name of the customer.
        lastName:
          type: string
        passportNo:
          type: string
        phoneNo:
          type: string
        email:
          type: string
          description: Email address of the customer.
        address:
          type: string
          description: Address of the customer.
      required:
        - id
        - firstName
        - lastName
        - email
        - address

    CustomerCreate:
      type: object
      properties:
        firstName:
          type: string
          description: Name of the customer.
        lastName:
          type: string
        passportNo:
          type: string
        phoneNo:
          type: string
        email:
          type: string
          description: Email address of the customer.
        address:
          type: string
          description: Address of the customer.
      required:
        - firstName
        - lastName
        - email
        - address

    CustomerUpdate:
      type: object
      properties:
        firstName:
          type: string
          description: Name of the customer.
        lastName:
          type: string
        passportNo:
          type: string
        phoneNo:
          type: string
        email:
          type: string
          description: Email address of the customer.
        address:
          type: string
          description: Address of the customer.
      required:
        - firstName
        - lastName
        - email
        - address

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        reason:
          type: string
        code:
          type: string
      required:
        - message